+++ python/mujoco/indexer_xmacro.h
@@ -23,6 +23,7 @@
   X( int,     actuator_, gaintype,     nu, 1                    ) \
   X( int,     actuator_, biastype,     nu, 1                    ) \
   X( int,     actuator_, trnid,        nu, 2                    ) \
+  X( int,     actuator_, actadr,       nu, 1                    ) \
   X( int,     actuator_, group,        nu, 1                    ) \
   X( mjtByte, actuator_, ctrllimited,  nu, 1                    ) \
   X( mjtByte, actuator_, forcelimited, nu, 1                    ) \